Register missing operators and make types header only.
