+++ src/torch_exports.cpp
@@ -44,6 +44,13 @@ void register_callables (DllInfo *dll)
   REGISTER_C_CALLABLE(operator_sexp_optional_double)
   REGISTER_C_CALLABLE(operator_sexp_optional_bool)
   REGISTER_C_CALLABLE(operator_sexp_optional_int64_t)
+  REGISTER_C_CALLABLE(operator_sexp_bool)
+  REGISTER_C_CALLABLE(operator_sexp_double)
+  REGISTER_C_CALLABLE(operator_sexp_int64_t)
+  REGISTER_C_CALLABLE(operator_sexp_optional_scalar)
+  REGISTER_C_CALLABLE(operator_sexp_optional_string)
+  REGISTER_C_CALLABLE(operator_sexp_optional_scalar_type)
+  REGISTER_C_CALLABLE(operator_sexp_optional_memory_format)
   
   REGISTER_C_CALLABLE(from_sexp_tensor)
   REGISTER_C_CALLABLE(from_sexp_optional_tensor)
@@ -85,6 +92,12 @@ void register_callables (DllInfo *dll)
   REGISTER_C_CALLABLE(from_sexp_optional_generator)
   REGISTER_C_CALLABLE(from_sexp_optional_double)
   REGISTER_C_CALLABLE(from_sexp_double)
+  REGISTER_C_CALLABLE(from_sexp_int64_t)
+  REGISTER_C_CALLABLE(from_sexp_scalar_type)
+  REGISTER_C_CALLABLE(from_sexp_optional_scalar)
+  REGISTER_C_CALLABLE(from_sexp_optional_string)
+  REGISTER_C_CALLABLE(from_sexp_optional_scalar_type)
+  REGISTER_C_CALLABLE(from_sexp_optional_memory_format)
   
   REGISTER_C_CALLABLE(delete_tensor)
   REGISTER_C_CALLABLE(delete_script_module)
@@ -142,6 +155,10 @@ void register_callables (DllInfo *dll)
   REGISTER_C_CALLABLE(delete_optional_dimname_list)
   REGISTER_C_CALLABLE(delete_optional_generator)
   REGISTER_C_CALLABLE(delete_optional_double)
+  REGISTER_C_CALLABLE(delete_optional_string)
+  REGISTER_C_CALLABLE(delete_optional_scalar)
+  REGISTER_C_CALLABLE(delete_optional_scalar_type)
+  REGISTER_C_CALLABLE(delete_optional_memory_format)
   
   REGISTER_C_CALLABLE(fixme_new_string)
 }
